Basic version that does JSON over HTTP.
This works, but it seems to be doing two HTTP calls at once each time,
which is not so great. It also does not start rsense for you, you need
to `rsense start` yourself first.

This implementation leverages auto-complete, which is also written by
Tomohiro Matsuyama (m2ym). His version of Rsense also contains the emacs
integration:

https://github.com/m2ym/rsense/blob/master/etc/rsense.el

But this one relies on a special output format for Emacs, an option that
has since been removed from Rsense.

I would say we take his version and merge it with what I have here to do
the communication over JSON. Shouldn't be too hard and then it should
work the way it did on the old version.